+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.1.40] - 2019-10-04
+## [0.1.41] - 2019-10-04
 ### Added
-- Quick sort loop
+- Summary loop algorithm
+- Summary recursion algorithm
 
 ## [0.1.30] - 2019-10-03
 ### Added